Rename dir to fix
